Add CardSearch component and injected the logic from CardSearchPage into it for a better structure
